ci-test-requirements.txt
aiohttp==3.8.4 ; python_version >= "3.9" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
argcomplete==2.1.1 ; python_version >= "3.9" and python_version < "4.0"
arpeggio==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.9" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.9" and python_version < "4.0"
black==23.1.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
cfgv==3.3.1 ; python_version >= "3.9" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
codecov==2.1.12 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_system == "Windows"
coverage==7.2.1 ; python_version >= "3.9" and python_version < "4.0"
coverage[toml]==7.2.1 ; python_version >= "3.9" and python_version < "4.0"
datamodel-code-generator==0.17.1 ; python_version >= "3.9" and python_version < "4.0"
dill==0.3.6 ; python_version >= "3.9" and python_version < "4.0"
distlib==0.3.6 ; python_version >= "3.9" and python_version < "4.0"
dnspython==2.3.0 ; python_version >= "3.9" and python_version < "4.0"
email-validator==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
exceptiongroup==1.1.1 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.92.0 ; python_version >= "3.9" and python_version < "4.0"
filelock==3.9.0 ; python_version >= "3.9" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.9" and python_version < "4.0"
future==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
gax-google-logging-v2==0.8.3 ; python_version >= "3.9" and python_version < "4.0"
gax-google-pubsub-v1==0.8.3 ; python_version >= "3.9" and python_version < "4.0"
gcloud==0.18.3 ; python_version >= "3.9" and python_version < "4.0"
genson==1.2.2 ; python_version >= "3.9" and python_version < "4.0"
google-gax==0.12.5 ; python_version >= "3.9" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version >= "3.9" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.9" and python_version < "4.0" and platform_machine == "aarch64" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "ppc64le" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "x86_64" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "amd64" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "AMD64" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "win32" or python_version >= "3.9" and python_version < "4.0" and platform_machine == "WIN32"
grpc-google-logging-v2==0.8.1 ; python_version >= "3.9" and python_version < "4.0"
grpc-google-pubsub-v1==0.8.1 ; python_version >= "3.9" and python_version < "4.0"
grpcio==1.51.3 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
httplib2==0.21.0 ; python_version >= "3.9" and python_version < "4.0"
identify==2.5.19 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4.0"
importlib-resources==5.12.0 ; python_version >= "3.9" and python_version < "4.0"
inflect==5.6.2 ; python_version >= "3.9" and python_version < "4.0"
iniconfig==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
isort==5.12.0 ; python_version >= "3.9" and python_version < "4.0"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
jsonschema==4.17.3 ; python_version >= "3.9" and python_version < "4.0"
jsonschema-spec==0.1.3 ; python_version >= "3.9" and python_version < "4.0"
lazy-object-proxy==1.9.0 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.9" and python_version < "4.0"
mypy-extensions==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
nodeenv==1.7.0 ; python_version >= "3.9" and python_version < "4.0"
oauth2client==4.1.3 ; python_version >= "3.9" and python_version < "4.0"
openapi-schema-validator==0.3.4 ; python_version >= "3.9" and python_version < "4.0"
openapi-spec-validator==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
parver==0.4 ; python_version >= "3.9" and python_version < "4.0"
pathable==0.4.3 ; python_version >= "3.9" and python_version < "4.0"
pathspec==0.11.0 ; python_version >= "3.9" and python_version < "4.0"
platformdirs==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
pluggy==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
ply==3.8 ; python_version >= "3.9" and python_version < "4.0"
prance==0.22.2.22.0 ; python_version >= "3.9" and python_version < "4.0"
pre-commit==3.1.1 ; python_version >= "3.9" and python_version < "4.0"
protobuf==4.22.1 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.9" and python_version < "4.0"
pulumi==3.57.1 ; python_version >= "3.9" and python_version < "4.0"
pulumi-gcp==6.51.0 ; python_version >= "3.9" and python_version < "4.0"
pulumi-kubernetes==3.24.1 ; python_version >= "3.9" and python_version < "4.0"
pulumi-random==4.12.0 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.6 ; python_version >= "3.9" and python_version < "4.0"
pydantic[email]==1.10.6 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
pyrsistent==0.19.3 ; python_version >= "3.9" and python_version < "4.0"
pysnooper==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
pytest==7.2.2 ; python_version >= "3.9" and python_version < "4.0"
pytest-asyncio==0.21.0 ; python_version >= "3.9" and python_version < "4.0"
pytest-cov==4.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
ruamel-yaml==0.17.21 ; python_version >= "3.9" and python_version < "4.0"
ruamel-yaml-clib==0.2.7 ; platform_python_implementation == "CPython" and python_version < "3.11" and python_version >= "3.9"
semver==2.13.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==67.6.0 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==2.0.5.post1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.25.0 ; python_version >= "3.9" and python_version < "4.0"
toml==0.10.2 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_full_version <= "3.11.0a6"
typed-ast==1.5.4 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4"
uvicorn==0.20.0 ; python_version >= "3.9" and python_version < "4.0"
virtualenv==20.20.0 ; python_version >= "3.9" and python_version < "4.0"
yarl==1.8.2 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "3.10"
